+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/runner/localRun/localsettings/FunctionLocalSettingsUtil.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
+ * Copyright 2018-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
  */
 
 package com.microsoft.azure.toolkit.intellij.legacy.function.runner.localRun.localsettings
@@ -21,6 +21,7 @@ object FunctionLocalSettingsUtil {
     private val jsonFormat = Json {
         decodeEnumsCaseInsensitive = true
         explicitNulls = false
+        ignoreUnknownKeys = true
     }
 
     fun readFunctionLocalSettings(publishableProject: PublishableProjectModel): FunctionLocalSettings? =